chore: reconcile Lotto vs lotto directories (#126)
the @gamble/lotto package was split between directories called `Lotto`
and `lotto`. On Mac's case-insensitive filesystem this doesn't seem to
cause any issues, but on Linux it would -- at least -- cause a
discrepancy in yarn.lock when running `yarn install`. This commit moves
all the files to the `Lotto` directory and gets rid of `lotto`.